Drop all tenant schemas before running db:dev_seed
When running this task we truncate all tables; however, doing so doesn't
execute the `after_destroy` callback which drops the generated schema.
That meant we could run into a situation where there are schemas in the
database with no associated tenant, leading to data inconsistencies.

So we're now destroying the tenants (alongside their schemas) before
truncating the rest of the database tables.