Allow different custom images per tenant
Note this only affects images which can also be customized using the
administration interface; other images like `avatar_admin.png` must be
the same for all tenants. I think this is good enough for now, since the
images that can't be different are the images that aren't customized
that often, and if there's a need to use different images in a certain
CONSUL installation, it can be achieved by changing the code which
renders that image so it uses `image_path_for`.

Since we're now using the same "tenants/#{Tenant.current_schema}"
subfolder pattern in three different places, we're refactoring the code
a bit. I'm not sure about the method name (`subfolder_path`) though.

Using this method for the current tenant in the `TenantDiskService`
class resulted in a `//` in the path, which is probably harmless but
very ugly and it also generates a different key than the one we got
until now. I've added an extra test to make sure that isn't the case.